Final Draft #1
The sending after response is not working; this will have to be
reimplemented. Sending text is working thus we will present ours stating
that it is sending text and we will later implement google API. We also
need to fix the text boxes so that text box appears before the list of
contacts. Let's finish this.